support start/stop iscsi container

+++ lib/lxc-manager/lxc-controller.rb
@@ -122,6 +122,7 @@ class LxcManager
 			template = File.join( config['template_dir'], distro.template )
 			repo_url = File.join( config['repo_url'], distro.id.to_s )
 
+			zfs_pool_zvol_path    = config['zfs_pool_zvol_path']
 			dir_pool_lxc_path     = config['dir_pool_lxc_path']
 			dir_pool_share_path   = config['dir_pool_share_path']
 			dir_export_lxc_path   = config['dir_export_lxc_path']
@@ -130,12 +131,14 @@ class LxcManager
 			dir_mount_distro_path = config['dir_mount_distro_path']
 			dir_mount_share_path  = config['dir_mount_share_path']
 			dir_root_dev_path     = config['dir_root_dev_path']
+			pool_zvol_path    = File.join( "/dev/zvol", zfs_pool_zvol_path, container.id.to_s )
 			pool_lxc_path     = File.join( dir_pool_lxc_path, container.id.to_s )
 			export_lxc_path   = File.join( dir_export_lxc_path, container.id.to_s )
 			mount_lxc_path    = File.join( dir_mount_lxc_path, container.id.to_s )
 			mount_distro_path = File.join( dir_mount_distro_path, distro.id.to_s )
 			root_dev_path     = File.join( dir_root_dev_path, container.id.to_s )
 
+			inter_host_network_interface_v4_address = "#{config['inter_host_network_interface_v4_address']}"
 			allowed_clients = "#{config['inter_host_network_v4_address']}/#{config['inter_host_network_v4_prefix']}"
 
 			target = Hash.new
@@ -147,61 +150,175 @@ class LxcManager
 
 			logger.debug "#{self}##{__method__}: " + "cli-agent start"
 			CliAgent.open( config['local_shell'] ){ |s|
+				targetcli_ls_success = false
+				targetcli_create_iblock_success = false
+				targetcli_create_iscsi_success = false
+				targetcli_set_attribute_success = false
+				targetcli_create_luns_success = false
+				targetcli_create_portals_success = false
+				targetcli_create_acls_success = false
 				jump_success = false
 				mkdir_root_dev_success = false
 				make_dev_files_success = false
+				iscsiadm_discovery_success = false
+				iscsiadm_node_success = false
+				iscsiadm_node_login_success = false
+				iscsiadm_session_success = false
+				mkdir_mount_lxc_path_success = false
+				mount_success = false
 				lxc_start_success = false
 
 				begin
+					if container.storage_type == LxcManager::Container::StorageType::ISCSI
+						ret = s.run "targetcli ls /"
+						if s.exit_status == 0
+							targetcli_ls_success = true
+						else
+							raise "Failed: targetcli ls: #{ret}"
+						end
+
+						ret = s.run "targetcli /backstores/iblock create name=#{container.id.to_s} dev=#{pool_zvol_path}"
+						if s.exit_status == 0
+							targetcli_create_iblock_success = true
+						else
+							raise "Failed: targetcli create iblock: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi create iqn.2016-11.com.example:#{container.id.to_s}"
+						if s.exit_status == 0
+							targetcli_create_iscsi_success = true
+						else
+							raise "Failed: targetcli create iscsi: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1 set attribute authentication=0"
+						if s.exit_status == 0
+							targetcli_set_attribute_success = true
+						else
+							raise "Failed: targetcli set attribute: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/luns create /backstores/iblock/#{container.id.to_s}"
+						if s.exit_status == 0
+							targetcli_create_luns_success = true
+						else
+							raise "Failed: targetcli create luns: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/portals create #{inter_host_network_interface_v4_address} 3260"
+						if s.exit_status == 0
+							targetcli_create_portals_success = true
+						else
+							raise "Failed: targetcli create portals: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/acls create iqn.2016-11.com.example:#{container.host.name}"
+						if s.exit_status == 0
+							targetcli_create_acls_success = true
+						else
+							raise "Failed: targetcli create acls: #{ret}"
+						end
+					end
+
 					s.jump( :ssh, target: target )
 					jump_success = true
 
-					ret = s.run "mkdir -p #{root_dev_path}"
-					if s.exit_status == 0
-						mkdir_root_dev_success = true
-					else
-						raise "Failed: mkdir -p: couldn't mkdir -p #{root_dev_path}"
-					end
-
-					command = Array.new
-					command.push "mknod -m 666  #{root_dev_path}/null c 1 3"
-					command.push "mknod -m 666  #{root_dev_path}/zero c 1 5"
-					command.push "mknod -m 666  #{root_dev_path}/random c 1 8"
-					command.push "mknod -m 666  #{root_dev_path}/urandom c 1 9"
-					command.push "mkdir -m 755  #{root_dev_path}/pts"
-					command.push "mkdir -m 1777 #{root_dev_path}/shm"
-					command.push "mknod -m 666  #{root_dev_path}/tty c 5 0"
-					command.push "mknod -m 666  #{root_dev_path}/tty0 c 4 0"
-					command.push "mknod -m 666  #{root_dev_path}/tty1 c 4 1"
-					command.push "mknod -m 666  #{root_dev_path}/tty2 c 4 2"
-					command.push "mknod -m 666  #{root_dev_path}/tty3 c 4 3"
-					command.push "mknod -m 666  #{root_dev_path}/tty4 c 4 4"
-					command.push "mknod -m 600  #{root_dev_path}/console c 5 1"
-					command.push "mknod -m 666  #{root_dev_path}/full c 1 7"
-					command.push "mknod -m 600  #{root_dev_path}/initctl p"
-					command.push "mknod -m 666  #{root_dev_path}/ptmx c 5 2"
-					command.each{ |c|
-						ret = s.run c
-						if s.exit_status != 0
-							raise "Failed: #{c}"
+					if container.storage_type == LxcManager::Container::StorageType::NFS
+						ret = s.run "mkdir -p #{root_dev_path}"
+						if s.exit_status == 0
+							mkdir_root_dev_success = true
+						else
+							raise "Failed: mkdir -p: couldn't mkdir -p #{root_dev_path}"
 						end
-					}
 
-					command_mknod_loop = String.new
-					command_mknod_loop += "for i in $(seq 0 255);"
-					command_mknod_loop += "do"
-					command_mknod_loop += "  if [ ! -b #{root_dev_path}/loop$i ];"
-					command_mknod_loop += "  then"
-					command_mknod_loop += "    /bin/mknod -m 0640 #{root_dev_path}/loop$i b 7 $i;"
-					command_mknod_loop += "    /bin/chown root:disk #{root_dev_path}/loop$i;"
-					command_mknod_loop += "  fi;"
-					command_mknod_loop += "done"
-					ret = s.run command_mknod_loop
-					if s.exit_status != 0
-						raise "Failed: #{command_mknod_loop}"
+						command = Array.new
+						command.push "mknod -m 666  #{root_dev_path}/null c 1 3"
+						command.push "mknod -m 666  #{root_dev_path}/zero c 1 5"
+						command.push "mknod -m 666  #{root_dev_path}/random c 1 8"
+						command.push "mknod -m 666  #{root_dev_path}/urandom c 1 9"
+						command.push "mkdir -m 755  #{root_dev_path}/pts"
+						command.push "mkdir -m 1777 #{root_dev_path}/shm"
+						command.push "mknod -m 666  #{root_dev_path}/tty c 5 0"
+						command.push "mknod -m 666  #{root_dev_path}/tty0 c 4 0"
+						command.push "mknod -m 666  #{root_dev_path}/tty1 c 4 1"
+						command.push "mknod -m 666  #{root_dev_path}/tty2 c 4 2"
+						command.push "mknod -m 666  #{root_dev_path}/tty3 c 4 3"
+						command.push "mknod -m 666  #{root_dev_path}/tty4 c 4 4"
+						command.push "mknod -m 600  #{root_dev_path}/console c 5 1"
+						command.push "mknod -m 666  #{root_dev_path}/full c 1 7"
+						command.push "mknod -m 600  #{root_dev_path}/initctl p"
+						command.push "mknod -m 666  #{root_dev_path}/ptmx c 5 2"
+						command.each{ |c|
+							ret = s.run c
+							if s.exit_status != 0
+								raise "Failed: #{c}"
+							end
+						}
+
+						command_mknod_loop = String.new
+						command_mknod_loop += "for i in $(seq 0 255);"
+						command_mknod_loop += "do"
+						command_mknod_loop += "  if [ ! -b #{root_dev_path}/loop$i ];"
+						command_mknod_loop += "  then"
+						command_mknod_loop += "    /bin/mknod -m 0640 #{root_dev_path}/loop$i b 7 $i;"
+						command_mknod_loop += "    /bin/chown root:disk #{root_dev_path}/loop$i;"
+						command_mknod_loop += "  fi;"
+						command_mknod_loop += "done"
+						ret = s.run command_mknod_loop
+						if s.exit_status != 0
+							raise "Failed: #{command_mknod_loop}"
+						end
+
+						make_dev_files_success = true
+					elsif container.storage_type == LxcManager::Container::StorageType::ISCSI
+						ret = s.run "iscsiadm -m discovery --type sendtargets --portal #{inter_host_network_interface_v4_address}"
+						if s.exit_status == 0
+							iscsiadm_discovery_success = true
+						else
+							raise "Failed: iscsiadm -m discovery: #{ret}"
+						end
+
+						ret = s.run "iscsiadm -m node"
+						if s.exit_status == 0
+							iscsiadm_node_success = true
+						else
+							raise "Failed: iscsiadm -m node: #{ret}"
+						end
+
+						ret = s.run "iscsiadm -m node -T iqn.2016-11.com.example:#{container.id.to_s} -p #{inter_host_network_interface_v4_address} --login"
+						if s.exit_status == 0
+							iscsiadm_node_login_success = true
+						else
+							raise "Failed: iscsiadm -m node --login: #{ret}"
+						end
+
+						ret = s.run "iscsiadm -m session"
+						if s.exit_status == 0
+							iscsiadm_session_success = true
+						else
+							raise "Failed: iscsiadm -m session"
+						end
+
+						ret = s.run "mkdir -p #{mount_lxc_path}"
+						if s.exit_status == 0
+							mkdir_mount_lxc_path_success = true
+						else
+							raise "Failed: mkdir -p: couldn't mkdir -p #{mount_lxc_path}"
+						end
+
+						ret = s.run "mountpoint -q #{mount_lxc_path}"
+						if s.exit_status != 0
+							disk_by_path = "/dev/disk/by-path/ip-#{inter_host_network_interface_v4_address}:3260-iscsi-iqn.2016-11.com.example:#{container.id.to_s}-lun-0"
+							ret = s.run "mount -t xfs #{disk_by_path} #{mount_lxc_path}"
+							if s.exit_status == 0
+								mount_success = true
+							else
+								raise "Failed: Mount: couldn't mount #{mount_zvol_path} to #{mount_lxc_path}"
+							end
+						else
+							mount_success = true
+						end
 					end
-
-					make_dev_files_success = true
 
 					ret = s.run "cat /proc/mounts | grep #{mount_lxc_path}"
 
@@ -220,12 +337,50 @@ class LxcManager
 					rescue
 					end
 				rescue
+					if mount_success
+						ret = s.run "umount -l #{mount_lxc_path}"
+					end
+
+					if mkdir_mount_lxc_path_success
+						ret = s.run "mountpoint -q #{mount_lxc_path}"
+						if s.exit_status != 0
+							ret = s.run "rm -rf #{mount_lxc_path}"
+						else
+							raise "Failed: Mount: couldn't mount #{mount_zvol_path} to #{mount_lxc_path}"
+						end
+					end
+
+					if iscsiadm_node_login_success
+						ret = s.run "iscsiadm -m node -T iqn.2016-11.com.example:#{container.id.to_s} -p #{inter_host_network_interface_v4_address} --logout"
+						ret = s.run "iscsiadm -m session"
+					end
+
 					if mkdir_root_dev_success
 						ret = s.run "rm -rf #{root_dev_path}"
 					end
 
 					if jump_success
 						s.jump( :exit )
+					end
+
+					if targetcli_create_acls_success
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/acls delete iqn.2016-11.com.example:#{container.host.name}"
+					end
+
+					if targetcli_create_portals_success
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/portals delete #{inter_host_network_interface_v4_address} 3260"
+					end
+
+					if targetcli_create_luns_success
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/luns delete /backstores/iblock/#{container.id.to_s}"
+					end
+
+					if targetcli_create_iscsi_success
+						ret = s.run "targetcli /iscsi delete iqn.2016-11.com.example:#{container.id.to_s}"
+					end
+
+					if targetcli_create_iblock_success
+						ret = s.run "targetcli /backstores/iblock delete name=#{container.id.to_s}"
 					end
 
 					raise
@@ -246,6 +401,7 @@ class LxcManager
 			template = File.join( config['template_dir'], distro.template )
 			repo_url = File.join( config['repo_url'], distro.id.to_s )
 
+			zfs_pool_zvol_path    = config['zfs_pool_zvol_path']
 			dir_pool_lxc_path     = config['dir_pool_lxc_path']
 			dir_pool_share_path   = config['dir_pool_share_path']
 			dir_export_lxc_path   = config['dir_export_lxc_path']
@@ -254,12 +410,14 @@ class LxcManager
 			dir_mount_distro_path = config['dir_mount_distro_path']
 			dir_mount_share_path  = config['dir_mount_share_path']
 			dir_root_dev_path     = config['dir_root_dev_path']
+			pool_zvol_path    = File.join( "/dev/zvol", zfs_pool_zvol_path, container.id.to_s )
 			pool_lxc_path     = File.join( dir_pool_lxc_path, container.id.to_s )
 			export_lxc_path   = File.join( dir_export_lxc_path, container.id.to_s )
 			mount_lxc_path    = File.join( dir_mount_lxc_path, container.id.to_s )
 			mount_distro_path = File.join( dir_mount_distro_path, distro.id.to_s )
 			root_dev_path     = File.join( dir_root_dev_path, container.id.to_s )
 
+			inter_host_network_interface_v4_address = "#{config['inter_host_network_interface_v4_address']}"
 			allowed_clients = "#{config['inter_host_network_v4_address']}/#{config['inter_host_network_v4_prefix']}"
 
 			target = Hash.new
@@ -276,9 +434,14 @@ class LxcManager
 				stop_complete_success = false
 				remove_dev_success = false
 				umount_mount_success = false
+				iscsiadm_node_logout_success = false
 				exit_success = false
-				uexport_success = false
-				umount_export_success = false
+				targetcli_delete_acls_success = false
+				targetcli_delete_portals_success = false
+				targetcli_delete_luns_success = false
+				targetcli_delete_iscsi_success = false
+				targetcli_delete_iblock_success = false
+				targetcli_ls_success = false
 
 				begin
 					s.jump( :ssh, target: target )
@@ -301,11 +464,13 @@ class LxcManager
 						raise "Failed: stop container: couldn't complete stop #{container.name} in 120 sec"
 					end
 
-					ret = s.run "rm -rf #{root_dev_path}"
-					if s.exit_status == 0
-						remove_dev_success = true
-					else
-						raise "Failed: remove /dev: couldn't remove /dev directory"
+					if container.storage_type == LxcManager::Container::StorageType::NFS
+						ret = s.run "rm -rf #{root_dev_path}"
+						if s.exit_status == 0
+							remove_dev_success = true
+						else
+							raise "Failed: remove /dev: couldn't remove /dev directory"
+						end
 					end
 
 					ret = s.run "mkdir -p #{mount_lxc_path}"
@@ -321,8 +486,96 @@ class LxcManager
 						umount_mount_success = true
 					end
 
+					if container.storage_type == LxcManager::Container::StorageType::ISCSI
+						ret = s.run "iscsiadm -m session"
+						if s.exit_status == 0
+							iscsiadm_session_success = true
+						else
+							raise "Failed: iscsiadm -m session"
+						end
+
+						ret = s.run "iscsiadm -m node -T iqn.2016-11.com.example:#{container.id.to_s} -p #{inter_host_network_interface_v4_address} --logout"
+						if s.exit_status == 0
+							iscsiadm_discovery_success = true
+						else
+							raise "Failed: iscsiadm -m node --logout: #{ret}"
+						end
+
+						ret = s.run "iscsiadm -m session"
+						if s.exit_status == 0
+							iscsiadm_session_success = true
+						else
+							raise "Failed: iscsiadm -m session"
+						end
+
+						ret = s.run "iscsiadm -m node"
+						if s.exit_status == 0
+							iscsiadm_node_success = true
+						else
+							raise "Failed: iscsiadm -m node: #{ret}"
+						end
+					end
+
 					s.jump( :exit )
 					exit_success = true
+
+					if container.storage_type == LxcManager::Container::StorageType::ISCSI
+						ret = s.run "targetcli ls /"
+						if s.exit_status == 0
+							targetcli_ls_success = true
+						else
+							raise "Failed: targetcli ls: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/acls delete iqn.2016-11.com.example:#{container.host.name}"
+						if s.exit_status == 0
+							targetcli_delete_acls_success = true
+						else
+							raise "Failed: targetcli delete acls: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/portals delete #{inter_host_network_interface_v4_address} 3260"
+						if s.exit_status == 0
+							targetcli_delete_portals_success = true
+						else
+							raise "Failed: targetcli delete portals: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1/luns delete /backstores/iblock/#{container.id.to_s}"
+						if s.exit_status == 0
+							targetcli_delete_luns_success = true
+						else
+							raise "Failed: targetcli delete luns: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi/iqn.2016-11.com.example:#{container.id.to_s}/tpg1 set attribute authentication=0"
+						if s.exit_status == 0
+							targetcli_set_attribute_success = true
+						else
+							raise "Failed: targetcli set attribute: #{ret}"
+						end
+
+						ret = s.run "targetcli /iscsi delete iqn.2016-11.com.example:#{container.id.to_s}"
+						if s.exit_status == 0
+							targetcli_delete_iscsi_success = true
+						else
+							raise "Failed: targetcli delete iscsi: #{ret}"
+						end
+
+						ret = s.run "targetcli /backstores/iblock delete name=#{container.id.to_s} dev=#{pool_zvol_path}"
+						if s.exit_status == 0
+							targetcli_delete_iblock_success = true
+						else
+							raise "Failed: targetcli delete iblock: #{ret}"
+						end
+
+						ret = s.run "targetcli ls /"
+						if s.exit_status == 0
+							targetcli_ls_success = true
+						else
+							raise "Failed: targetcli ls: #{ret}"
+						end
+					end
 				rescue
 					raise
 				end
